# backend-service/internal/handler/laptop_handler.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"

	// Sesuaikan "skripsi-backend" dengan nama module di file go.mod kamu
	"github.com/vk878930/skripsi-backend/entity"
	"github.com/vk878930/skripsi-backend/internal/usecase"
)

type LaptopHandler struct {
	laptopUsecase usecase.LaptopUsecase
}

// NewLaptopHandler adalah constructor
func NewLaptopHandler(r *gin.Engine, us usecase.LaptopUsecase) {
	handler := &LaptopHandler{
		laptopUsecase: us,
	}

	// Mendaftarkan routing endpoint
	r.POST("/api/estimasi", handler.EstimasiHarga)
	r.GET("/api/estimasi/riwayat", handler.AmbilRiwayatPrediksi)
	r.POST("/api/estimasi/evaluate", handler.EvaluasiModel)
	r.POST("/api/estimasi/update_k", handler.UpdateK)
}

// EstimasiHarga godoc
// @Summary Calculate laptop price estimation
// @Description Send laptop specification to get price estimation from ML service
// @Tags Estimasi
// @Accept json
// @Produce json
// @Param spek body entity.SpesifikasiLaptop true "Laptop Specification"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /estimasi [post]
func (h *LaptopHandler) EstimasiHarga(c *gin.Context) {
	var spek entity.SpesifikasiLaptop

	// 1. Validasi input JSON
	if err := c.ShouldBindJSON(&spek); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Format spesifikasi tidak valid"})
		return
	}

	// 2. Lempar ke Usecase untuk diproses (Handler tidak perlu tahu cara kerja Python)
	hasil, err := h.laptopUsecase.DapatkanEstimasiHarga(spek)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	// 3. Kembalikan hasil ke user
	c.JSON(http.StatusOK, gin.H{
		"pesan": "Berhasil memproses via arsitektur clean",
		"data":  hasil,
	})
}

// AmbilRiwayatPrediksi godoc
// @Summary Get all price estimation history
// @Description Retrieve the history of all price estimations requested
// @Tags Estimasi
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /estimasi/riwayat [get]
func (h *LaptopHandler) AmbilRiwayatPrediksi(c *gin.Context) {
	// Panggil fungsi dari usecase (Sesuaikan nama fungsinya dengan yang ada di usecase kamu)
	data, err := h.laptopUsecase.DapatkanRiwayat() 
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Gagal mengambil riwayat prediksi"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": data})
}

func (h *LaptopHandler) EvaluasiModel(c *gin.Context) {
	var body struct {
		K int `json:"k"`
	}
	if err := c.ShouldBindJSON(&body); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Format tidak valid, k harus berupa angka"})
		return
	}

	hasil, err := h.laptopUsecase.EvaluasiModel(body.K)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": hasil})
}

func (h *LaptopHandler) UpdateK(c *gin.Context) {
	var body struct {
		K int `json:"k"`
	}
	if err := c.ShouldBindJSON(&body); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Format tidak valid, k harus berupa angka"})
		return
	}

	hasil, err := h.laptopUsecase.UpdateK(body.K)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": hasil})
}